+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [Unreleased]
+  - 💄 [@mixte/snippets] 优化 `Lottery` 的 `selectCard` 展示卡片的 z 值计算逻辑
 
 ## [v3.2.0-beta.3]
   - 📅 2025-01-08